+++ calc_testify_test/calc_testify_test.go
@@ -8,17 +8,17 @@ import (
 
 func TestAdd(t *testing.T) {
 	result := Add(1, 2)
-	Equal(t, int64(3), result)
+	Equal(t, 3, result)
 }
 
 func TestSubstract(t *testing.T) {
 	result := Substract(5, 3)
-	Equal(t, int64(2), result)
+	Equal(t, 2, result)
 }
 
 func TestMultiply(t *testing.T) {
 	result := Multiply(5, 6)
-	Equal(t, int64(30), result)
+	Equal(t, 30, result)
 }
 
 func TestDivide(t *testing.T) {
